NTUEEPLUS-44. omit some content of recruitment and recommendation (#28)
* create lower level Block

* finish show more in recommendation

* finish show more in recommendation

* rename Block and Post

* move inline style into css
