+++ src/views/components/career/recruitment/Recruitment.css
@@ -20,13 +20,19 @@
   padding: 0%;
   border-radius: 5% !important;
 }
-.recrucontent{
-  padding:0% 3% 5% 3%;
+.recrucontent {
+  padding: 0% 3% 5% 3%;
 }
-.fw-semibold{
+.fw-semibold {
   font-size: 1.9rem !important;
   margin-top: 5%;
 }
-hr{
+hr {
   margin-top: 0;
+}
+.recrucontent > button {
+  font-size: 1vw;
+  color: blue;
+  border: none;
+  background: none;
 }
